+++ src/fidesops/schemas/user.py
@@ -1,8 +1,0 @@
-from fideslib.schemas.base_class import BaseSchema
-
-
-class UserPasswordReset(BaseSchema):
-    """Contains both old and new passwords when resetting a password"""
-
-    old_password: str
-    new_password: str