+++ tests/models/test_client.py
@@ -1,9 +1,9 @@
+from fideslib.cryptography.cryptographic_util import hash_with_salt
 from fideslib.models.client import ClientDetail
 from sqlalchemy.orm import Session
 
 from fidesops.api.v1.scope_registry import SCOPE_REGISTRY
 from fidesops.core.config import config
-from fidesops.util.cryptographic_util import hash_with_salt
 
 
 class TestClientModel: